Lift Push Knock Out is a project to show how to use Lift with knockout.js
For simplicity's sake, we'll be using Twitter's bootstrap in conjunction
with jQuery and knockout.js. All we want to do is build a nice, simple
chat client. We're not going to be complicating it with too many
features; initially, all we'll be looking at is a simple chat
interaction, but mediated entirely through a Lift comet.

I'll elaborate a little bit on that last part. Typically, Lift
interactions are defined in a snippet somewhere in the snippet package,
bound on the client by a CSS transform, and interactions with it happen
within the snippet package. Occasionally, the only thing the snippet
does is send a message to an intermediary actor, which then talks to the
comet.

At OpenStudy, we use a slightly different interaction model, and this
project will be demonstrating that model. The key difference is that we
bind the vast majority of our UI via knockout, so all form handling
functions and field names are also sent down via comet and bound via
knockout. This introduces a couple of small thorns that we need to deal
with in terms of bringing our handling in life with some of Lift's
assumptions, but they're fairly trivial. We'll look at them in a bit.

Starting Point
--------------

The project starts off from the basis of the lift_blank template in the
main Lift 2.4 download at http://liftweb.net/download . We strip out
some extraneous files, add Twitter bootstrap, the latest jQuery, and
knockout.js, and make the main package com.withoutincident instead of
the default code that is used in the Lift template. Then, we're ready to
start!