First off, we're going to set up a quick skeleton of what we're going to be looking at. We're not super worried about seeing online users at the moment; all we want to implement is basic chat. So our template contains just enough to display chat and give us a field to type our chat into. We use a bit of bootstrap magic and some simple styles of our own to fix the message and post buttons to the bottom of the screen. We also style the ordered list tag that we'll be using to display chats so that it has the traditional chat content layout.
Now we have our base view to work with; our next step is the backend
logic. Note that we won't be hooking anything up with the client side
yet; we're just going to set up the comet interactions that will post
and receive messages. We'll do the actual form interactions and knockout
bindings after all the infrastructure is set up on both ends.