The backend chat setup is actually super-easy. We have a basic ChatActor, modeled mostly after Simply Lift's ChatServer. Paired to it is a per-client ChatComet, which takes care of posting and receiving messages and handling changes in usernames. Each user gets a default username using Lift's built-in randon number generator function, nextNum, which we expose by importing net.liftweb.util.Helpers._ .
Last but not least, every new client gets the last message that was sent
(rather than all messages ever, for our purposes we stick to the last
one). We leave stubs for now for how we will call down into the client;
we'll be filling those in after we implement the JavaScript side of
the interactions, so that we know what we want to call into from the
server.